pyproject.toml
[project]
name = "stac-fastapi.indexed"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "asgi-correlation-id~=4.3.3",
    "async-lru~=2.0.5",
    "brotli_asgi~=1.4.0",
    "duckdb~=1.2.2",
    "orjson",
    "pydantic-settings",  # don't specify a version, determined by stac-pydantic
    "pygeofilter[backend-native]==0.2.1",
    "pyjwt~=2.8.0", "pytz",
    "shapely~=2.0.4",
    "stac-fastapi.api==3.0.1",
    "stac-fastapi.extensions==3.0.1",
    "stac-fastapi.types==3.0.1",
    "stac-pydantic~=3.1.0",
    "stac-index",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.1.0",
    "pytest~=8.2.2",
    "requests>=2.32.3",
]
server = [
    "uvicorn~=0.30.1",
]
lambda = [
    "mangum~=0.17.0",
]
test = [
    "pytest~=8.2.2",
]
iac = [
    "aws-cdk-lib~=2.189.0",
]

[tool.pytest.ini_options]
pythonpath = ["./src"]
norecursedirs = [
    "tests/with_environment/integration_tests",
    "tests/with_environment/smoke_tests",
]

[tool.uv.sources]
stac-index = { workspace = true }


[tool.uv.workspace]
members = ["packages/*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src/"]
include = ["stac_fastapi.geoparquet"]
